Added UIFillBar. Added Assessment recap. Assessment results are now tied to the pillars view.

+++ Assets/_core/_scripts/core/Data/CategoryData.cs
@@ -6,7 +6,7 @@ namespace Ieedo
     public class CategoryData
     {
         public CategoryID ID;
-        public int AssessmentValue;
+        public float AssessmentValue;
 
         public override string ToString()
         {
